requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Faker==12.0.1
Markdown==3.4.1
mixer==7.2.2
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
